pyproject.toml
[tool.poetry]
name = "mlight"
version = "0.1.0"
description = "Light client for python with custom RS485 protocol"
authors = ["Daniel Hnyk <hnykda@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
paho-mqtt = "*"
typer = "*"
pyserial = "*"

[tool.poetry.dev-dependencies]
black = "*"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
